Linux: use an int pointer for the size argument to getsockopt
getsockopt expects an int pointer as the size parameter, both in
the kernel version and our replacement compat inline function, which
causes warnings because we give it a size_t pointer.

Use an int variable instead.

Reviewed-on: http://gerrit.openafs.org/2417
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/rx/LINUX/rx_knet.c
@@ -159,7 +159,7 @@ osi_NetSend(osi_socket sop, struct sockaddr_in *to, struct iovec *iovec,
     int code;
 #ifdef ADAPT_PMTU
     int sockerr;
-    size_t esize;
+    int esize;
 
     while (1) {
 	sockerr=0;
@@ -211,7 +211,7 @@ osi_NetReceive(osi_socket so, struct sockaddr_in *from, struct iovec *iov,
     int code;
 #ifdef ADAPT_PMTU
     int sockerr;
-    size_t esize;
+    int esize;
 #endif
     struct iovec tmpvec[RX_MAXWVECS + 2];
     struct socket *sop = (struct socket *)so;
